requirements.txt
Django~=3.1.4
numpy~=1.19.4
dlib~=19.21.0
requests~=2.25.0
Pillow~=8.0.1
opencv-python~=4.4.0.46
twilio~=6.49.0
eciespy
face_recognition
pycrypto